+++ spec/fixtures/Project[With]Special{chars}in*path?/Project[With]Special{chars}in*path?/main.m
@@ -1,0 +1,18 @@
+//
+//  main.m
+//  Project[With]Special{chars}in*path?
+//
+//  Created by Olivier Halligon on 22/11/2014.
+//  Copyright (c) 2014 CocoaPods. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+#import "AppDelegate.h"
+
+int main(int argc, char *argv[])
+{
+  @autoreleasepool {
+      return UIApplicationMain(argc, argv, nil, NSStringFromClass([AppDelegate class]));
+  }
+}