+++ spec/fixtures/Project[With]Special{chars}in*path?/TestRunner/AppDelegate.h
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.h
+//  TestRunner
+//
+//  Created by Eloy Durán on 3/27/12.
+//  Copyright (c) 2012 __MyCompanyName__. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AppDelegate : UIResponder <UIApplicationDelegate>
+
+@property (strong, nonatomic) UIWindow *window;
+
+@end