+++ CHANGELOG.md
@@ -32,6 +32,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
   [Samuel Giddins](https://github.com/segiddins)
   [#2876](https://github.com/CocoaPods/CocoaPods/issues/2876)
 
+* Fix issues when trying to discover the xcodeproj automatically
+  but the path contains special chars (`[`,`]`,`{`,`}`,`*`,`?`)
+  [Olivier Halligon](https://github.com/AliSoftware)
+  [#2852](https://github.com/CocoaPods/CocoaPods/issues/2852)
 
 ## 0.35.0